pyproject.toml
[project]
name = "dociq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.5.4",
    "pydantic-settings>=2.10.1",
    "pypdf>=5.9.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
